Add solution to 13b
Uhh, so this was a fun one. Somehow I still remembered the Chinese
Remainder Theorem from my number theory class and figured it could
probably apply here, but wasn't sure if the numbers would be too big to
store in a 64-bit long.

Took like a day to remember how CRT worked and then figure out how to
algorithmically solve it, then took a few more hours to get the types
and signedness correct. I think the bus IDs have to be coprime for
there to be a proper solution and so that CRT can be applied here, but
TBH I also peeked at the input and saw that all the bus IDs were prime.