Add solution to 20a
This one...took a really long time (compared to the other problems). Had
a false start on how the tiles were supposed to be arranged since I
didn't fully read the problem (assumed the solution was just 3x3 like
the example, lol), but after figuring out a different approach for
arranging the tiles, it was smooth sailing.

Also the longest program and most complex in terms of pieces of
abstraction used. Took 2 enums, 3 structs, 1 class, 2 type aliases, and
a magic array of arrays.